mobile view

+++ unicorn.sh
@@ -1,18 +1,8 @@
 #!/bin/sh  
 # 
 
-if test -z "$2"
-then
-    PORT=3456
-else
-    PORT=$2
-fi
-if test -z "$3"
-then
-    ENVIRONMENT="production"
-else
-    ENVIRONMENT="$3"
-fi
+PORT=$(test -z "$2" && echo "3456" || echo "$2")
+ENVIRONMENT=$(test -z "$3" && echo "production" || echo "$3")
 
 echo "port: ${PORT} environment: ${ENVIRONMENT}"
 UNICORN=unicorn  
@@ -22,18 +12,27 @@ case "$1" in
     start)  
         test -d log || mkdir log
         test -d tmp || mkdir -p tmp/pids
-        echo "stop watch_dog.sh"
-        /bin/sh watch_dog.sh stop
+
+        echo "stop nohup"
+        /bin/sh nohup.sh stop
+
         echo "start unicorn"
         bundle exec ${UNICORN} -c ${CONFIG_FILE} -p ${PORT} -E ${ENVIRONMENT} -D  
-        echo "start watch_dog.sh"
-        /bin/sh watch_dog.sh start
+        
+        echo "start nohup"
+        /bin/sh nohup.sh start
         ;;  
     stop)  
         echo "stop unicorn"
-        kill -QUIT `cat tmp/pids/unicorn.pid`  
-        echo "stop watch_dog.sh"
-        /bin/sh watch_dog.sh stop
+        if test -f tmp/pids/unicorn.pid 
+        then
+            kill -QUIT `cat tmp/pids/unicorn.pid`  
+        else
+            echo "[WARNGIN]: unicorn.pid not exist."
+        fi
+
+        echo "stop nohup"
+        /bin/sh nohup.sh stop
         ;;  
     restart|force-reload)  
         kill -USR2 `cat tmp/pids/unicorn.pid`  
@@ -44,6 +43,7 @@ case "$1" in
         echo "RACK_ENV=${ENVIRONMENT} bundle exec rake agent:deploy"
         echo "RACK_ENV=${ENVIRONMENT} bundle exec rake agent:check"
         echo "RACK_ENV=${ENVIRONMENT} bundle exec rake agent:main"
+        echo "RACK_ENV=${ENVIRONMENT} bundle exec rake remote:deploy"
         ;;
     *)  
         echo "Usage: $SCRIPTNAME {start|stop|restart|force-reload|rake}" >&2
